Split `PathRegistry` into its own file.
